fix(data-model): split CompositeSubFieldLabels getter into its own file
CI's twenty(max-consts-per-file) rule allows only one top-level const
per file, and I had both the COMPOSITE_SUB_FIELD_LABELS map and the
getCompositeSubFieldLabel helper in CompositeSubFieldLabels.ts. Moved
the helper into a sibling utils/getCompositeSubFieldLabel.ts. Also
merged a duplicate twenty-shared/utils import in
SettingsObjectIndexesSection.tsx flagged by no-duplicate-imports.
